moves get_laue_group_num to utils.py
* adds a phil param for setting it
