+++ simtbx/diffBragg/hopper_utils.py
@@ -634,6 +634,9 @@ class DataModeler:
         if not self.params.fix.diffuse_gamma or not self.params.fix.diffuse_sigma:
             assert self.params.use_diffuse_models
         self.SIM.D.use_diffuse = self.params.use_diffuse_models
+        if self.params.use_diffuse_models and self.params.dissue_use_laue_symm:
+            assert self.params.space_group is not None
+            self.SIM.D.laue_group_num = utils.get_laue_group_num(self.params.space_group)  # TODO this can also be retrieved from crystal model if params.space_group is None
         self.SIM.D.gamma_miller_units = self.params.gamma_miller_units
         self.SIM.isotropic_diffuse_gamma = self.params.isotropic.diffuse_gamma
         self.SIM.isotropic_diffuse_sigma = self.params.isotropic.diffuse_sigma
@@ -1792,24 +1795,3 @@ def print_profile(stats, timed_methods):
             line = fp[l-1][:-1]
             PROFILE_LOGGER.warning("%5d%14.2f%9.2f%s" % (l, timespent[i_l]*unit*1e3, frac_t, line))
 
-def get_laue_group_number(sg_symbol=None):
-    """ get laue group number from space group symbol """
-    import cctbx.sgtbx
-    import re
-    if sg_symbol is None:
-        laue_sym="P-1"
-    else:
-        g = cctbx.sgtbx.space_group_info(sg_symbol).group()
-        if g.laue_group_type() not in ["2/m", "-3m"]:
-            laue_sym = "P{}".format(g.laue_group_type())
-        else:
-            pg = str(g.build_derived_patterson_group().info().symbol_and_number())
-            lc = re.sub(r'\([^)]*\)', '', pg[2:]).replace(" ", "")
-            if pg[0] == "R":
-                lc = lc.replace(":H", "1")
-            laue_sym = "P{}".format(lc)
-            
-
-    hm_symbols=['P-1','P112/m','P12/m1','P2/m11','Pmmm','P4/m','P4/mmm','P-3','P-3m1','P-31m','P6/m','P6/mmm','Pm-3','Pm-3m']
-    lgs = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14]
-    return lgs[hm_symbols.index(laue_sym)]
