+++ simtbx/diffBragg/src/util.h
@@ -105,8 +105,8 @@ struct flags{
 };
 
 struct crystal{
-    int laue_group_num=5;
-    int stencil_size=0; 
+    int laue_group_num=1;
+    int stencil_size=0;
     Eigen::Matrix3d anisoG;
     std::vector<Eigen::Matrix3d,Eigen::aligned_allocator<Eigen::Matrix3d> > dG_dgamma;
     std::vector<Eigen::Matrix3d,Eigen::aligned_allocator<Eigen::Matrix3d> > dU_dsigma;
